pyproject.toml
[project]
name = "barebones-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.6.0",
]

[project.scripts]
barebones-agent = "barebones_agent.agent:agent_loop"

[tool.setuptools.packages.find]
where = ["src"]
